Improve the interactivity of the section on holes in Naturals (#1089)
As discussed in https://github.com/plfa/plfa.github.io/issues/1088, I
suggest this small change to improve the interactivity of the file on
Naturals a bit, and hopefully remove the confusion that I at least had
with it.

As I explain in the text, the only price to pay is that we have to use
`_+'_` (plus prime) instead of `_+_`, because it was already defined in
the file.

---------

Co-authored-by: Christian Jauvin <cjauvin@explorance.com>
Co-authored-by: Wen Kokke <wenkokke@users.noreply.github.com>