TableOfContents.md: mention PLFArend as a Derived work (#1106)
Hello! I have been working on rewriting the PLFA in Arend. @wadler
suggested mentioning it as a derived work, thank you for this
opportunity!
https://agda.zulipchat.com/#narrow/channel/238741-general
I have added the attribution to the README.md of my project as required
by the CC-BY-4.0. Once again, sorry for not doing it in the first place.
No bad intentions, I just considered this work to be a personal project
that very few people would be interested in.
Also, I have exacted my work to a separate repo (thanks for the PLFArend
name idea!). This way it is clearly separated from the original book.